Add react-router setup with nested dashboard
