+++ src/components/Sidebar.jsx
@@ -6,9 +6,9 @@ export default function Sidebar() {
       <h2 className="text-2xl font-bold">TeCuido</h2>
       <nav className="flex flex-col gap-3">
         <Link to="/dashboard" className="hover:underline">Dashboard</Link>
-        <Link to="/medications" className="hover:underline">Medications</Link>
-        <Link to="/appointments" className="hover:underline">Appointments</Link>
-        <Link to="/chatbot" className="hover:underline">Chatbot</Link>
+        <Link to="/dashboard/medications" className="hover:underline">Medications</Link>
+        <Link to="/dashboard/appointments" className="hover:underline">Appointments</Link>
+        <Link to="/dashboard/chatbot" className="hover:underline">Chatbot</Link>
       </nav>
     </div>
   );
