+++ src/pages/Dashboard.jsx
@@ -1,3 +1,4 @@
+import { Outlet } from 'react-router-dom';
 import Sidebar from '../components/Sidebar';
 
 export default function Dashboard() {
@@ -5,7 +6,7 @@ export default function Dashboard() {
     <div className="flex">
       <Sidebar />
       <div className="flex-1 p-6">
-        <h1 className="text-2xl font-bold">📋 Dashboard</h1>
+        <Outlet />
       </div>
     </div>
   );
